requirements.txt
langchain>=0.1.0
langchain-anthropic>=0.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
bcrypt>=4.0.0
